refactor: seperate routers, dependencies, etc.
