+++ web/.env.example
@@ -6,4 +6,4 @@ SHOPIFY_APP_SECRET=d31xxxxxxxxxxxxxa7
 SHOPIFY_APP_SCOPES="read_products,write_metaobject_definitions"
 
 # ngrok url
-SHOPIFY_APP_URL="https://2560bc69926a.ngrok-free.app"
+SHOPIFY_APP_URL="https://d768bf47be7c.ngrok-free.app"
