+++ web/routers/__init__.py
@@ -1,0 +1,10 @@
+"""
+Initialize routers package.
+Export all routers for use in main application.
+"""
+
+from .auth import router as auth_router
+from .sync import router as sync_router
+from .api import router as api_router
+
+__all__ = ['auth_router', 'sync_router', 'api_router']
